Fix working state with JOBS_PER_FORK enabled
Normally, Resque records the state of the worker in redis in the parent
process, forks a child process for the job, and then cleans up the
working state when the child returns. With JOBS_PER_FORK enabled, we
need to update the worker's state in redis from the child process for
each job.